Move time.js to /src/constants/
