+++ atd-vzv/src/views/summary/queries/socrataQueries.js
@@ -1,4 +1,9 @@
-import { thisYear, lastYear, today, todayMonthYear } from "../helpers/time";
+import {
+  thisYear,
+  lastYear,
+  today,
+  todayMonthYear
+} from "../../../constants/time";
 
 export const crashEndpointUrl = `https://data.austintexas.gov/resource/y2wy-tgr5.json`;
 export const demographicsEndpointUrl = `https://data.austintexas.gov/resource/xecs-rpy9.json`;
